+++ lib/td_lm/permissions/permissions.ex
@@ -5,33 +5,31 @@ defmodule TdLm.Permissions do
 
   import Ecto.Query, warn: false
 
-  alias TdLm.Accounts.User
-
-  @permission_resolver Application.get_env(:td_lm, :permission_resolver)
+  alias TdLm.Auth.Claims
 
   @doc """
-  Check if user has a permission in a domain.
+  Check if authenticated user has a permission in a domain.
 
   ## Examples
 
-      iex> authorized?(%User{}, :create, "business_concept", 12)
+      iex> authorized?(%Claims{}, :create, "business_concept", 12)
       false
 
   """
-  def authorized?(%User{jti: jti}, permission, resource_type, id) do
-    @permission_resolver.has_permission?(jti, permission, resource_type, id)
+  def authorized?(%Claims{jti: jti}, permission, resource_type, id) do
+    TdCache.Permissions.has_permission?(jti, permission, resource_type, id)
   end
 
   @doc """
-  Check if user has a any permission in a domain.
+  Check if authenticated user has a any permission in a domain.
 
   ## Examples
 
-      iex> authorized_any?(%User{}, [:create, :delete], "business_concept", 12)
+      iex> authorized_any?(%Claims{}, [:create, :delete], "business_concept", 12)
       false
 
   """
-  def authorized_any?(%User{jti: jti}, permissions, resource_type, id) do
-    @permission_resolver.has_any_permission?(jti, permissions, resource_type, id)
+  def authorized_any?(%Claims{jti: jti}, permissions, resource_type, id) do
+    TdCache.Permissions.has_any_permission?(jti, permissions, resource_type, id)
   end
 end