+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changed
 
 - [TD-3083] Enrich cached ingest attributes
+- [TD-3163] Auth tokens now include `role` claim instead of `is_admin` flag
+- [TD-3182] Allow to use redis with password
 
 ## [4.11.0] 2021-01-11